feat: Add job name/id to job wait/log output and improve timestamp handling in the 'logs' command. (#725)
I received some offline feedback for the previous change adding these commands.

I have added job name/id to the output of the 'wait' command, as if you use the default job there is no indication of which job you are waiting on.

I've improved the timestamp handling in the logs command, it now returns ISO8601 and there is a '--timezone local' switch which will print out in local timezone:

Signed-off-by: Preston Tamkin <845970+prestomation@users.noreply.github.com>