test: Isolate integration tests from the user's config
I found that running the integration tests would fail depending on the
state of my config file in ~/.deadline/config. This modifies the
integration test suite so every test function gets an isolated config
file. With this change, the tests run reliably independent of how I set
various options in the config.

Signed-off-by: Mark <399551+mwiebe@users.noreply.github.com>